==> jupyter_notebooks/1.1_stats_b_data.ipynb <==
# 1.1 Player Stats Data Collection
---
*By Ihza Gonzales*

This notebook aims to create a dataframe of names of active non pitching baseball players in use to get the batting statistics of the players from the website Savantbaseball.mlb.com. The statistics of the baseball players from the years 2018 - 2021 will be collected.

## Import Libraries
---

import pandas as pd
import requests
from bs4 import BeautifulSoup
import time

## Import Data for Player Names and ID's
---

player_id = pd.read_csv('../data/og_datasets/SFBB Player ID Map - PLAYERIDMAP.csv')

player_id.head()

## Import List of all Active non-Pitching Players From Rotowire.com
---
*Obtained data has active column but not up-to data*

batters = pd.read_csv('../data/og_datasets/mlb-player-stats-Batters.csv')
batters.head()

# Want to get players with a minimum of 50 at-bats
batters = batters[batters['AB']> 50]

### Specify which Columns and Rows Needed

#Only get players that are not pitchers
player_id = player_id[player_id['POS'] != 'P']

mlb_id = player_id[['MLBID', 'FIRSTNAME', 'LASTNAME', 'ACTIVE', 'PLAYERNAME']]

mlb_id.head(3)

#### Only Want Active Players. Merge both Datasets to get mlbid of all active non pitching players

active = mlb_id.merge(batters, how = 'inner', left_on = 'PLAYERNAME', right_on = 'Player')
active.head(5)

#### Change ID to Integer

active['MLBID'] = active['MLBID'].astype(int)

active.shape

#### Drop any Duplicate IDs

active.drop_duplicates(subset=['MLBID'], keep='last', inplace = True)

# Make sure no duplicates
active[active.duplicated(['MLBID'])]

#### Drop column playername

active.drop(columns = 'PLAYERNAME', inplace = True)
active.head(3)

#### Save Dataframe with Id and names

active.to_csv('../data/og_datasets/mlb_players_bat.csv')

## Function for Collecting Stats from Baseballsavant.mlb.com
----

def get_stats(mlbid, first_name, last_name):
    """
    This function collects the gamelog stats of every 
    game for that season for the specified player from the arguments.
    
    Returns csv file named after the player of their stats to specified location.
    """
    years = ['2018', '2019', '2020', '2021']
    
    for year in years:
        #This is the url to the website
        base_url = 'https://baseballsavant.mlb.com/savant-player/'

        #This string will be used to specifiy the player
        player_name = first_name.lower() +'-'+last_name.lower()+'-'+str(mlbid)

        #Url for the page with the stats
        url = base_url + player_name + '?stats=gamelogs-r-hitting-mlb&season=' + year

        #Requests for the page
        res = requests.get(url)

        if res.status_code != 200:
            raise Exception('API response: {}'.format(res.status_code))   
            #Modified from https://pypi.org/project/ratelimit/

        else:

            soup = BeautifulSoup(res.content, 'lxml')

            player_stats = []

            try:

                #Find the table with desired stats
                table = soup.find('div', {'id':['gamelogs-mlb']})

                #Finds all the columns needed
                for row in table.find('tbody').find_all('tr'):

                    td_tags = row.find_all('td')
                    for index, td in enumerate(td_tags):
                        stats = {}
                        stats['date'] = td_tags[0].text.strip()
                        stats['PA'] = td_tags[3].text.strip()
                        stats['AB'] = td_tags[4].text.strip()
                        stats['R'] = td_tags[5].text.strip()
                        stats['H'] = td_tags[6].text.strip()
                        stats['2B'] = td_tags[7].text.strip()
                        stats['3B'] = td_tags[8].text.strip()
                        stats['HR'] = td_tags[9].text.strip()
                        stats['RBI'] = td_tags[10].text.strip()
                        stats['BB'] = td_tags[11].text.strip()
                        stats['SO'] = td_tags[12].text.strip()
                        stats['AVG'] = td_tags[16].text.strip()
                        stats['OBP'] = td_tags[17].text.strip()
                        stats['SLG'] = td_tags[18].text.strip()
                        stats['OPS'] = td_tags[19].text.strip()

                    #Appends the row of stats to the list
                    player_stats.append(stats)

                #Creates data frame of all stats
                df = pd.DataFrame(player_stats)

                #Saves Dataframe to a file with player name
                df.to_csv(f'../data/og_players_bat/{first_name}-{last_name}-{mlbid}-{year}.csv')

                time.sleep(1) #suspends execution for 1 second to prevent too many requests
                #inspired from https://realpython.com/python-sleep/

            except (AttributeError, IndexError):
                pass


## Get Stats for All Active Batters
---

%%time
for index, row in active.iterrows():
    
    mlbid = row['MLBID']
    first = row['FIRSTNAME']
    last = row['LASTNAME']

    get_stats(mlbid, first, last)
    
# Copied from https://stackoverflow.com/questions/16476924/how-to-iterate-over-rows-in-a-dataframe-in-pandas

## Recap
---
Saved a csv file of active batters names with ID's and current 2021 regular season stats. This file will be used to get names of players. The data collected from BaseballSavant.com are the stats of every game of active non-pitching players from 2018 - 2021. This will be used for the VAR model for forecasting.
