==> jupyter_notebooks/1.3_salary_data.ipynb <==
# 1.2 Player Salary Data Collection
---
*By Ihza Gonzales*

This notebook aims to collect the current salaries of active players from spotrac.com and append that data to the active players dataset

## Import Libraries
---

import pandas as pd
import requests
from bs4 import BeautifulSoup

## Salaries for Batters
---
### Requests the webpage

url = 'https://www.spotrac.com/mlb/rankings/2021/salary/batters/'

data = {
    'ajax': 'true',
    'mobile': 'false'
}

res = requests.post(url, data=data)

print(res.status_code)

#Copied from https://stackoverflow.com/questions/63376961/beautifulsoup-only-returing-100-elements

soup = BeautifulSoup(res.content, 'lxml')

### Find the names and salaries from the table and create a dataframe

table = soup.find('table')

batters = []

for row in table.find('tbody').find_all('tr'):
    batter = {}
    batter['player'] = row.find('h3').text.strip()
    batter['salary'] = row.find('span', {'class':'info'}).text.strip()
    
    batters.append(batter)
    
salary = pd.DataFrame(batters)

### Read in dataset of current batters

df = pd.read_csv('../data/og_datasets/mlb_players_bat.csv').drop('Unnamed: 0', axis = 1)
df.head()

### Merge the salary dataset to current batters

df = df.merge(salary, how = 'inner', left_on = 'Player', right_on = 'player')
df.head()

### Drop any duplicates

df.drop_duplicates(subset=['MLBID'], keep='first', inplace = True)

# Make sure no duplicates
df[df.duplicated(['MLBID'])]

### Drop one of the columns that was merged on

df.drop(columns = 'player', inplace = True)
df.head(3)

### Save updated dataset with salary

df.to_csv('../data/mlb_players_bat.csv')

## Salaries for Pitchers
---
### Requests the webpage

url = 'https://www.spotrac.com/mlb/rankings/2021/salary/pitching/'

data = {
    'ajax': 'true',
    'mobile': 'false'
}

res = requests.post(url, data=data)

print(res.status_code)

#Copied from https://stackoverflow.com/questions/63376961/beautifulsoup-only-returing-100-elements

soup = BeautifulSoup(res.content, 'lxml')

### Find the names and salaries from the table and create a dataframe

table = soup.find('table')

pitchers = []

for row in table.find('tbody').find_all('tr'):
    pitcher = {}
    pitcher['player'] = row.find('h3').text.strip()
    pitcher['salary'] = row.find('span', {'class':'info'}).text.strip()
    
    pitchers.append(pitcher)
    
salary = pd.DataFrame(pitchers)

### Read in dataset of current pitchers

df = pd.read_csv('../data/og_datasets/mlb_players_pitch.csv').drop('Unnamed: 0', axis = 1)
df.head()

### Merge the salary dataset to current pitchers

df = df.merge(salary, how = 'inner', left_on = 'Player', right_on = 'player')
df.head()

### Drop any duplicates

df.drop_duplicates(subset=['MLBID'], keep='first', inplace = True)

# Make sure no duplicates
df[df.duplicated(['MLBID'])]

### Drop one of the columns that was merged on

df.drop(columns = 'player', inplace = True)
df.head(3)

### Save updated dataset with salary

df.to_csv('../data/mlb_players_pitch.csv')

## Recap
---
Collected the 2021 salaries of current batters and pitchers and merged the data into the current batters and pitchers dataset. This will be used for the regression model for predicting salaries.
