==> jupyter_notebooks/2.1_active_players_cleaning.ipynb <==
# Active Players Cleaning
---
*By Ihza Gonzales*

This notebook aims to clean the current batters and pitchers with salary datasets. Removing unnecessary columns and checking for null. The stats must be the same stats gathered for other datasets. 

## Import Libraries
---

import pandas as pd
import numpy as np
pd.set_option('display.max_columns', None)

## Batters
---

df = pd.read_csv('../data/mlb_players_bat.csv').drop(['Unnamed: 0'], axis = 1)
df.head()

### Remove unwanted columns

df = df.drop(columns = ['ACTIVE', 'SB', 'HBP', 'SH', 'SF', 'CS'])
df.head()

### Check for Nulls

df.isnull().sum()

### Save clean dataset

df.to_csv('../data/mlb_players_bat.csv')

## Pitchers
---

df = pd.read_csv('../data/mlb_players_pitch.csv').drop(['Unnamed: 0'], axis = 1)
df.head()

### Remove unwanted columns

df = df[['MLBID', 'FIRSTNAME', 'LASTNAME', 'Player', 'Team', 'Age', 'W', 'L', 'ERA', 'IP', 'H', 'ER', 'HR', 'BB', 'K', 'WHIP', 'salary']]
df.head()

### Check for nulls

df.isnull().sum()

### Save clean dataset

df.to_csv('../data/mlb_players_pitch.csv')

## Recap
---
The active players dataset for pitchers and batters were cleaned and will be used throughout the project to check for player names and ID.
