==> jupyter_notebooks/2.4_salary_b_cleaning.ipynb <==
# Batter Salary Data Cleaning
---
*By Ihza Gonzales*

This notebook aims to clean the historical data of stats and salaries acquired from SeanLahmen.com. The data for salaries is from 1985-2016. 

## Import Libraries
---

import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)

## Import Historical Salary Dataset
---

salary = pd.read_csv('../data/lahman_database/Salaries.csv')
salary.head()

## Import Historical Batting Stats
---

bat = pd.read_csv('../data/lahman_database/Batting.csv')
bat.head()

### Only get stats after 1985 
*This is because salary dataset starts at 1985*

bat = bat[bat['yearID']>1985]
bat = bat[bat['AB']> 50]
bat.head()

### Create new column for batting average

bat['AVG'] = round((bat['H'] / bat['AB']), 3)
bat.head()

### Create new column for on-base percentage

numerator = bat['H'] + bat['BB'] + bat['HBP']
plate = bat['AB'] + bat['BB'] + bat['HBP'] + bat['SF']
bat['OBP'] = round((numerator / plate), 3)
bat.head()

### Create new column for slugging percentage

first = bat['H'] - (bat['2B'] + bat['3B'] + bat['HR'])

bat['SLG'] = round(((first + (2 * bat['2B']) + (3 * bat['3B']) + (4 * bat['HR'])) / bat['AB']), 3)

bat.head()

### Create new column for on-base plus slugging percentage

bat['OPS'] = bat['OBP'] + bat['SLG']

bat.head()

### Drop unwanted columns

bat.drop(columns = ['GIDP', 'IBB', 'stint', 'SB', 'HBP', 'SH', 'SF', 'CS'], inplace = True)

bat.head()

### Merge salary and batting dataset

df = bat.merge(salary, how = 'inner', left_on = ['playerID', 'yearID'], right_on = ['playerID', 'yearID'])
df.head()

### Drop repeated columns from merge

df.drop(columns = ['teamID_y', 'lgID_y'], inplace = True)
df.head()

df.isnull().sum()

### Save clean dataset

df.to_csv('../data/past_salaries_bat.csv')

## Recap
---
Merged and cleaned the dataset for use in the regression model. Added new columns to match the stats scrapped for batters.
