==> jupyter_notebooks/2.5_salary_p_cleaning.ipynb <==
# Pitcher Salary Data Cleaning
---
*By Ihza Gonzales*

This notebook aims to clean the historical data of stats and salaries acquired from SeanLahmen.com. The data for salaries is from 1985-2016. 

## Import Libraries
---

import numpy as np
import pandas as pd
pd.set_option('display.max_columns', None)

## Import Historical Salary Dataset
---

salary = pd.read_csv('../data/lahman_database/Salaries.csv')
salary.head()

## Import Historical Pitching Stats
---

pitch = pd.read_csv('../data/lahman_database/Pitching.csv')
pitch.head()

### Only get stats after 1985
*This is because salary dataset starts at 1985*

pitch = pitch[pitch['yearID']>1985]
pitch.head()

### Create new column for innings pitches

pitch['IP'] = round(pitch['IPouts']/3, 1)
pitch.head()

### Create new column for whip

pitch['WHIP'] = round((pitch['BB'] + pitch['H']) / (pitch['IPouts']/3), 2)
pitch.head()

### Drop unwanted columns

pitch = pitch[['playerID', 'yearID', 'W', 'L', 'ERA', 'IP', 'H', 'ER', 'HR', 'BB', 'SO', 'WHIP']]

pitch.head()

### Merge salary and pitching dataset

df = pitch.merge(salary, how = 'inner', left_on = ['playerID', 'yearID'], right_on = ['playerID', 'yearID'])
df.head()

### Drop repeated columns from merge

df.drop(columns = ['teamID', 'lgID'], inplace = True)
df.head()

### Check for nulls

df.isnull().sum()

df = df.dropna()

### Save clean dataset

df.to_csv('../data/past_salaries_pitch.csv')

## Recap
---
Merged and cleaned the dataset for use in the regression model. Added new columns to match the stats scrapped for pitchers.
