==> jupyter_notebooks/4.1_time_series_b.ipynb <==
# Batter Time Series Analysis
---
*By Ihza Gonzales*

This notebook aims to create a time series model for batters to forecast their stats. The data being used here is the stats scraped from baseballsavant. 

## Import Libraries
---

import pandas as pd
import numpy as np
import pickle

import matplotlib.pyplot as plt
import seaborn as sns

from statsmodels.tsa.api import VAR
from statsmodels.tsa.stattools import adfuller  
from sklearn.model_selection import train_test_split

import warnings
warnings.filterwarnings('ignore')

## Functions Implemented
---

def var_model(first, last, mlbid):
    
    """
    Function to create a VAR model for each individual player
    Returns a csv with forcasted stats
    """
    
    base_path = '../data/clean_players_bat/'
    
    #This string will be used to specifiy the player
    player_name = first + '-' + last + '-' + str(mlbid)
    
    #Full path to file
    file_path = base_path + player_name + '.csv'

    try:
        df = pd.read_csv(file_path)

        train, test = train_test_split(df, test_size = 0.15, shuffle = False)

        model = VAR(train)

        ts_model = model.fit()

        forecast = ts_model.forecast(train.values, len(test))

        player_stats = []
        
        for time in forecast:
            stats = {}

            stats['PA'] = round(time[0], 0)
            stats['AB'] = round(time[1], 0)
            stats['R'] = round(time[2], 0)
            stats['H'] = round(time[3], 0)
            stats['2B'] = round(time[4], 0)
            stats['3B'] = round(time[5], 0)
            stats['HR'] = round(time[6], 0)
            stats['RBI'] = round(time[7], 0)
            stats['BB'] = round(time[8], 0)
            stats['SO'] = round(time[9], 0)
            stats['AVG'] = round(time[10], 3)
            stats['OBP'] = round(time[11], 3)
            stats['SLG'] = round(time[12], 3)
            stats['OPS'] = round(time[13], 3)

            #Appends the row of stats to the list
            player_stats.append(stats)

        #Creates data frame of all stats
        df = pd.DataFrame(player_stats)

        #Saves Dataframe to a file with player name
        df.to_csv(f'../data/predictions_bat/{first}-{last}-{mlbid}.csv', index = False)
        
    except:
        print(f'{first} {last} something went wrong')

def average_forecast (first, last, mlbid):
    """
    Function to take the forecasted stats and average them for use in predicting the salary
    """
    base_path = '../data/predictions_bat/'
    
    #This string will be used to specifiy the player
    player_name = first + '-' + last + '-' + str(mlbid)
    
    #Full path to file
    file_path = base_path + player_name + '.csv'
    
    df = pd.read_csv(file_path)

    stats = {
        "name" : f'{first} {last}',
        "AB" : round(df['AB'].sum(), 0),
        "R" : round(df['R'].sum(), 0),
        "H" : round(df['H'].sum(), 0),
        "2B" : round(df['2B'].sum(), 0),
        "3B" : round(df['3B'].sum(), 0),
        "HR" : round(df['HR'].sum(), 0),
        "RBI" : round(df['RBI'].sum(), 0),
        "BB" : round(df['BB'].sum(), 0),
        "SO" : round(df['SO'].sum(), 0),
        "AVG" : round(df['AVG'].mean(), 3),
        "OBP" : round(df['OBP'].mean(), 3),
        "SLG" : round(df['SLG'].mean(), 3),
        "OPS" : round(df['OPS'].mean(), 3)
    }
    
    return stats

## Import Dataset of Active Batters
---

players = pd.read_csv('../data/mlb_players_bat.csv').drop('Unnamed: 0', axis = 1)
players.head()

## Run VAR Model for All Batters
---

%%time
for index, row in players.iterrows():
    
    first = row['FIRSTNAME']
    last = row['LASTNAME']
    mlbid = row['MLBID']
    
    var_model(first, last, mlbid)
    
print('Models Finished')
    
# Copied from https://stackoverflow.com/questions/16476924/how-to-iterate-over-rows-in-a-dataframe-in-pandas

## Take the Average and Count of Forecasted Stats to Use in Predicting Salaries
---

%%time
forecasts = []
for index, row in players.iterrows():
    
    first = row['FIRSTNAME']
    last = row['LASTNAME']
    mlbid = row['MLBID']
    
    forecasts.append(average_forecast(first, last, mlbid))
    
#Creates data frame of all stats
df = pd.DataFrame(forecasts)

#Saves Dataframe to a file with player name
df.to_csv(f'../data/player_forecasts_bat.csv')
    
#Save as pickle file
with open('../pickles/forecast_bat.pkl', 'wb') as pickle_out:
    pickle_out = pickle.dump(df, pickle_out)
    
print('Finished')

## EDA on Results
---

df = pd.read_csv('../data/clean_players_bat/Mookie-Betts-605141.csv')

train, test = train_test_split(df, test_size = 0.15, shuffle = False)

model = VAR(train)

ts_model = model.fit()

forecast = ts_model.forecast(train.values, len(test))

player_stats = []

for time in forecast:
    stats = {}

    stats['PA_f'] = round(time[0], 0)
    stats['AB_f'] = round(time[1], 0)
    stats['R_f'] = round(time[2], 0)
    stats['H_f'] = round(time[3], 0)
    stats['2B_f'] = round(time[4], 0)
    stats['3B_f'] = round(time[5], 0)
    stats['HR_f'] = round(time[6], 0)
    stats['RBI_f'] = round(time[7], 0)
    stats['BB_f'] = round(time[8], 0)
    stats['SO_f'] = round(time[9], 0)
    stats['AVG_f'] = round(time[10], 3)
    stats['OBP_f'] = round(time[11], 3)
    stats['SLG_f'] = round(time[12], 3)
    stats['OPS_f'] = round(time[13], 3)

    #Appends the row of stats to the list
    player_stats.append(stats)

#Creates data frame of all stats
df_results = pd.DataFrame(player_stats)

def forecast_accuracy(forecast, actual):
    rmse = np.mean((forecast - actual)**2)**.5  # RMSE
    return({'rmse':rmse})

print('Forecast Accuracy of: PA')
accuracy_prod = forecast_accuracy(df_results['PA_f'].values, test['PA'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))

print('\nForecast Accuracy of: AB')
accuracy_prod = forecast_accuracy(df_results['AB_f'].values, test['AB'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))

print('\nForecast Accuracy of: R')
accuracy_prod = forecast_accuracy(df_results['R_f'].values, test['R'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))

print('\nForecast Accuracy of: H')
accuracy_prod = forecast_accuracy(df_results['H_f'].values, test['H'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))

print('\nForecast Accuracy of: 2B')
accuracy_prod = forecast_accuracy(df_results['2B_f'].values, test['2B'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))

print('\nForecast Accuracy of: 3B')
accuracy_prod = forecast_accuracy(df_results['3B_f'].values, test['3B'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))

print('\nForecast Accuracy of: HR')
accuracy_prod = forecast_accuracy(df_results['HR_f'].values, test['HR'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))

print('\nForecast Accuracy of: RBI')
accuracy_prod = forecast_accuracy(df_results['RBI_f'].values, test['RBI'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))
    
print('\nForecast Accuracy of: BB')
accuracy_prod = forecast_accuracy(df_results['BB_f'].values, test['BB'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))
    
print('\nForecast Accuracy of: SO')
accuracy_prod = forecast_accuracy(df_results['SO_f'].values, test['SO'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))
    
print('\nForecast Accuracy of: AVG')
accuracy_prod = forecast_accuracy(df_results['AVG_f'].values, test['AVG'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))
    
print('\nForecast Accuracy of: OBP')
accuracy_prod = forecast_accuracy(df_results['OBP_f'].values, test['OBP'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))
    
print('\nForecast Accuracy of: SLG')
accuracy_prod = forecast_accuracy(df_results['SLG_f'].values, test['SLG'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))
    
print('\nForecast Accuracy of: OPS')
accuracy_prod = forecast_accuracy(df_results['OPS_f'].values, test['OPS'])
for k, v in accuracy_prod.items():
    print(k, ': ', round(v,4))
    
#Modified from https://www.machinelearningplus.com/time-series/vector-autoregression-examples-python/

## Recap
---
A VAR model was ran and used to forcast about a third of the season. Then the forcasted results were counted and averaged to give estimated stats for the upcoming season. The stats will be used for the app and to predict their salary.
