==> jupyter_notebooks/7.1_recommender_b.ipynb <==
# Batter Recommender System
---
*By Ihza Gonzales*

Using Current stats, salaries, and clusters, a recommender system was built using cosine distances.

## Import Libraries
---

import pandas as pd
import numpy as np 
from scipy import sparse
import pickle

from sklearn.metrics.pairwise import pairwise_distances, cosine_distances, cosine_similarity

## Import Dataset
---

df = pd.read_csv('../data/clusters_bat.csv').drop('Unnamed: 0', axis = 1)
df.head()

df.drop(columns = ['MLBID', 'FIRSTNAME', 'LASTNAME', 'Team', 'Pos', 'Age', 'G'], inplace = True)
df.head()

## Create Pivot Table
---

pivot = pd.pivot_table(df, index = 'Player')
pivot.head()

sparse_pivot = sparse.csr_matrix(pivot.fillna(0))

recommender = cosine_distances(sparse_pivot)   

recommender_df = pd.DataFrame(recommender, columns=pivot.index, index=pivot.index)
recommender_df.head()

## Test Recommender
---

recommender_df['Mookie Betts'].sort_values()[1:11]

## Save CSV and Pickle Files of Recommendation
---

recommender_df.to_csv('../data/recommender_bat.csv', index = False)

with open('../pickles/recommendation_bat.pkl', 'wb') as pickle_out:
    pickle_out = pickle.dump(recommender_df, pickle_out)

## Recap
---
Create a recommendation system from current stats, salaries and clusters for each batter. This will be used in the app
